backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
pydantic>=2.11.0
pydantic-settings>=2.7.0
httpx>=0.28.1
openai>=1.75.0
tavily-python>=0.5.0
